requirements.txt
pillow
scikit-image
numpy

# Development dependencies
-e .  # This assumes your project is in the current directory
black
flit